+++ packages/amplify-sdk/CHANGELOG.md
@@ -6,12 +6,17 @@
  * feat: Updated `appcd-fs` to add support for applying parent directory ownership when being
    executed as sudo. ([APIGOV-19102](https://jira.axway.com/browse/APIGOV-19102))
  * feat: Added list of teams to find org info.
+ * fix(switch): Added `isPlatformTooling` flag to properly handle logging out of a platform tooling
+   account. ([APIGOV-19370](https://jira.axway.com/browse/APIGOV-19370))
  * fix: Validate team argument for user add, list, and update commands.
    ([APIGOV-19216](https://jira.axway.com/browse/APIGOV-19216))
  * fix(teams): Allow team user management commands to case insensitive match team name or guid and
    match user by email address or guid.
    ([APIGOV-19105](https://jira.axway.com/browse/APIGOV-19105))
  * fix: Parse activity and usage dates using current locale.
+ * fix: Fixed platform URL reference in switch org logic.
+ * chore: Removed child organizations from organization info. Platform has deprecated parent/child
+   organizations. This also solves an issue where users of child orgs cannot view the parent org.
 
 # v2.1.5 (May 11, 2021)